# packages/contracts/openapi/v1.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIMS Frontend-Backend Contract",
    "version": "1.0.0",
    "description": "Canonical V1 API contract for gemma-policy-simulator. This OpenAPI artifact is the single source of truth for shared request/response shapes."
  },
  "servers": [
    {
      "url": "http://localhost:8000/api/v1"
    }
  ],
...
  "paths": {
    "/simulations": {
      "post": {
        "tags": [
          "simulations"
        ],
        "operationId": "createSimulation",
        "summary": "Create a new simulation draft",
        "description": "Creates a pending simulation draft. For dataset=nemotron_usa, any unsupported filter key must be rejected with 400 and error.code=UNSUPPORTED_FILTER. Rejection is strict: no partial acceptance and no silent key dropping.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSimulationRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Draft created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateSimulationEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Validation or unsupported filter error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "simulations"
        ],
        "operationId": "listSimulations",
        "summary": "List simulations",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/SimulationStatus"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "schema": {
              "type": "string",
              "example": "created_at:desc"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Simulation history",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SimulationListEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/simulations/{id}": {
      "delete": {
        "tags": [
          "simulations"
        ],
        "operationId": "deleteSimulation",
        "summary": "Delete simulation",
        "parameters": [
          {
            "$ref": "#/components/parameters/SimulationIdPath"
          }
        ],
        "responses": {
          "200": {
            "description": "Deleted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteSimulationEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
    "/simulations/{id}/clarifications/generate": {
      "post": {
        "tags": [
          "clarifications"
        ],
        "operationId": "generateClarification",
        "summary": "Generate next clarification question",
        "description": "Generates one clarification question via local Gemma runtime. Clarification generation is draft-only and does not change simulation lifecycle status.",
        "parameters": [
          {
            "$ref": "#/components/parameters/SimulationIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateClarificationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clarification generated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationQuestionEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Simulation not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle conflict for non-pending simulation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Gemma/Ollama runtime failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
    "/clarifications/{clarification_id}/answer": {
      "post": {
        "tags": [
          "clarifications"
        ],
        "operationId": "answerClarification",
        "summary": "Answer clarification question",
        "description": "Processes clarification answer via local Gemma runtime, updates refined policy text, and optionally opens the next clarification turn. Clarification flow is pending-only with a hard max of 3 turns.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ClarificationIdPath"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ClarificationAnswerRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Clarification answer processed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClarificationAnswerEnvelope"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "404": {
            "description": "Simulation or active clarification not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "409": {
            "description": "Lifecycle conflict for non-pending simulation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          },
          "500": {
            "description": "Gemma/Ollama runtime failure",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "SimulationIdPath": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
      "ClarificationIdPath": {
        "name": "clarification_id",
        "in": "path",
        "required": true,
        "schema": {
          "type": "string"
        }
      },
...
    },
    "schemas": {
      "SimulationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "running",
          "completed",
          "failed"
        ]
      },
...
      "ClarificationStatus": {
        "type": "string",
        "enum": [
          "none",
          "in_progress",
          "resolved"
        ]
      },
      "ClarificationQuestionState": {
        "type": "string",
        "enum": [
          "open",
          "resolved"
        ]
      },
      "Meta": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "request_id": {
            "type": "string"
          },
          "device_profile": {
            "type": "string"
          },
          "total": {
            "type": "integer"
          },
          "page": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          }
        }
      },
      "ErrorObject": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "ErrorEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "null"
          },
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "FilterSet": {
        "type": "object",
        "description": "For dataset=nemotron_usa, supported filter keys are states, age_range, sex, marital_status, education_level, occupation. Unsupported keys income_brackets, household_income, ethnicity must trigger 400 with error.code=UNSUPPORTED_FILTER.",
        "additionalProperties": false,
        "properties": {
          "states": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "age_range": {
            "type": "array",
            "items": {
              "type": "integer"
            },
            "minItems": 2,
            "maxItems": 2
          },
          "sex": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "marital_status": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "education_level": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "occupation": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "CreateSimulationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "title",
          "policy_text",
          "dataset",
          "sample_size"
        ],
        "properties": {
          "title": {
            "type": "string"
          },
          "policy_text": {
            "type": "string"
          },
          "dataset": {
            "type": "string"
          },
          "sample_size": {
            "type": "integer",
            "minimum": 20,
            "maximum": 2000
          },
          "filters": {
            "$ref": "#/components/schemas/FilterSet"
          }
        }
      },
      "SimulationDraft": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "policy_text",
          "status",
          "dataset",
          "sample_size",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "policy_text": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationStatus"
          },
          "dataset": {
            "type": "string"
          },
          "sample_size": {
            "type": "integer"
          },
          "filters": {
            "$ref": "#/components/schemas/FilterSet"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "CreateSimulationEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/SimulationDraft"
          },
          "error": {
            "type": "null"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
...
      "SimulationListItem": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "title",
          "status",
          "sample_size",
          "mean_approval",
          "created_at",
          "completed_at"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "title": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/SimulationStatus"
          },
          "sample_size": {
            "type": "integer"
          },
          "mean_approval": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "SimulationListEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SimulationListItem"
            }
          },
          "error": {
            "type": "null"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "DeleteSimulationData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "id",
          "deleted"
        ],
        "properties": {
          "id": {
            "type": "string"
          },
          "deleted": {
            "type": "boolean"
          }
        }
      },
      "DeleteSimulationEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/DeleteSimulationData"
          },
          "error": {
            "type": "null"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "GenerateClarificationRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "focus"
        ],
        "properties": {
          "focus": {
            "type": "string"
          }
        }
      },
      "ClarificationQuestionData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "clarification_id",
          "simulation_id",
          "question_text",
          "rationale",
          "status",
          "turn_index"
        ],
        "properties": {
          "clarification_id": {
            "type": "string"
          },
          "simulation_id": {
            "type": "string"
          },
          "question_text": {
            "type": "string"
          },
          "rationale": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/ClarificationQuestionState"
          },
          "turn_index": {
            "type": "integer",
            "minimum": 1
          }
        }
      },
      "ClarificationQuestionEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ClarificationQuestionData"
          },
          "error": {
            "type": "null"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
      "ClarificationAnswerRequest": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "simulation_id",
          "user_response"
        ],
        "properties": {
          "simulation_id": {
            "type": "string"
          },
          "user_response": {
            "type": "string"
          }
        }
      },
      "ClarificationAnswerData": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "simulation_id",
          "clarification_status",
          "refined_policy_text"
        ],
        "properties": {
          "simulation_id": {
            "type": "string"
          },
          "clarification_status": {
            "$ref": "#/components/schemas/ClarificationStatus"
          },
          "refined_policy_text": {
            "type": "string"
          },
          "next_clarification_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "next_question_text": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "ClarificationAnswerEnvelope": {
        "type": "object",
        "additionalProperties": false,
        "required": [
          "data",
          "error",
          "meta"
        ],
        "properties": {
          "data": {
            "$ref": "#/components/schemas/ClarificationAnswerData"
          },
          "error": {
            "type": "null"
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
...
    }
  }
}